Initial import of some JFPatch files and the WebSocket client.
This import grabs a bunch of random JFPatch files that I had lying
around so that people can at least see what they look like. They're
not checked for any sort of usefulness, and they don't really provide
much in the way of functionality anyhow.

Also imports the websockets client which will handle the communications
with the server.